Win32 projects: change the assembly target platform from x86 to AnyCPU for Visual Studio 2022 compatibility (#5)
Visual Studio 2022 is based on the x64 platform. Consequently, MSBuild is executed as a 64-bit process and therefore cannot load assemblies that were only compiled for 32-bit platform. This commit sets the target platform to AnyCPU so that the assemblies can be loaded in a 32-bit environment (VS2010 to VS2019) as well as in a 64-bit environment (VS2022).